rename: groupbuyingchat, groupbuyingwrite 에 속해있는 파일 이름 수정
