+++ app/src/main/java/com/umc/one_person_households_platform/view/groupbuyingchat/GroupBuyingChatMoreWriterFragment.kt
@@ -1,15 +1,15 @@
-package com.umc.one_person_households_platform.view.groupbuyingwrite
+package com.umc.one_person_households_platform.view.groupbuyingchat
 
 import android.os.Bundle
+import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-import com.umc.one_person_households_platform.databinding.FragmentGroupBuyingCategoryBinding
+import com.umc.one_person_households_platform.databinding.FragmentGroupBuyingChatMoreWriterBinding
 
-class GroupBuyingCategoryFragment : BottomSheetDialogFragment() {
+class GroupBuyingChatMoreWriterFragment : Fragment() {
 
-    private var _binding: FragmentGroupBuyingCategoryBinding? = null
+    private var _binding: FragmentGroupBuyingChatMoreWriterBinding? = null
     private val binding get() = _binding!!
 
     override fun onCreateView(
@@ -17,7 +17,7 @@ class GroupBuyingCategoryFragment : BottomSheetDialogFragment() {
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        _binding = FragmentGroupBuyingCategoryBinding.inflate(inflater, container, false)
+        _binding = FragmentGroupBuyingChatMoreWriterBinding.inflate(inflater, container, false)
         return binding.root
     }
