+++ app/src/main/res/layout/fragment_group_buying_chat_more_writer.xml
