+++ app/src/main/res/layout/fragment_group_buying_write_category.xml
@@ -10,8 +10,7 @@
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@drawable/bg_group_buying_rounded"
-        tools:context=".GroupBuyingCategoryFragment">
+        android:background="@drawable/bg_group_buying_rounded">
 
         <TextView
             android:id="@+id/tv_description"
