+++ app/src/main/res/layout/fragment_group_buying_write_count.xml
